refactor blockmeta/odin
