+++ blockmeta/odin/api.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
-from flask import request
 import json
+from utils import responses
 from flask_restful import Resource, reqparse
-from blockmeta.utils import util
 from manager import OdinManager
 from tools import flags
 
@@ -30,17 +29,17 @@ class OdinAPI(Resource):
         argument_list = [ arg for arg in arguments.split('/') if arg != ""]
         print len(argument_list)
         if len(argument_list) < 1 or argument_list[0] != self.ppk:
-            return util.wrap_ordin_response(400, uri)
+            return responses.wrap_ordin_response(400, uri)
         if len(argument_list) == 1:
-            return util.hello_world_response(uri)
+            return responses.hello_world_response(uri)
         elif len(argument_list) == 2:
-            return util.parse_next_response(uri)
+            return responses.parse_next_response(uri)
         elif len(argument_list) == 3:
             found = self.manager.handle_odin(argument_list[1], argument_list[2])
             if found:
-                return util.wrap_ordin_response(200, uri, found)
+                return responses.wrap_ordin_response(200, uri, found)
             else:
-                return util.wrap_ordin_response(404, uri)
+                return responses.wrap_ordin_response(404, uri)
         else:
             pass
         return None
